prova1/api/requirements.txt
flask
Flask-SQLAlchemy
requests
psycopg2-binary